server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.40.0",
]
